requirements.txt
requests
click
python-dateutil
rich
